requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0